@@ -3,16 +3,21 @@ name = starling-beancount
 
 [options]
 packages = starling_beancount
-python_requires = >= 3.6
+python_requires = >= 3.9
 install_requires =
     beancount
     httpx
-    pyyaml
+    PyYAML
     typer
 
-[options.entry_points]
-console_scripts =
-    starling = starling_beancount.extractor:cli
+[options.extras_require]
+dev =
+    black
+    flake8
+    isort
+    mypy
+    types-PyYAML
+    beancount-stubs
 
 [tool.black]
 line-length = 88
@@ -23,3 +28,17 @@ exclude = .git,.hg,.mypy_cache,.tox,.venv,_build,buck-out,build,dist
 max-line-length = 88
 exclude = .tox,.eggs,ci/templates,build,dist, __init__.py
 ignore=E741,F403,E265,W504,E226,W503,E501,E203
+
+[tool:isort]
+profile=black
+skip = .tox,.eggs,ci/templates,build,dist
+
+[mypy]
+disallow_untyped_defs = True
+disallow_any_unimported = True
+no_implicit_optional = True
+check_untyped_defs = True
+warn_return_any = True
+warn_unused_ignores = True
+show_error_codes = True
+warn_unreachable = True
